refactor(contracts): consolidate Track/TrackMetadata/SearchResult/TrackPayload
Establish canonical cross-process types in @shiranami/contracts and collapse
the duplicated copies onto them:

- TrackMetadata: dedupe the 4 desktop copies (metadata-service, scan-utility-host,
  scan-utility, preload/types) + the web electron.d.ts mirror onto one
  domain/media.ts shape.
- DisplayTrack: add the renderer display projection to domain/track.ts; the
  renderer stores/types Track now re-exports it (the orphaned DB-mirror Track in
  contracts is now the live canonical wire shape).
- SearchResult: dedupe utils/ytdlp-spawn + preload/api/downloader onto contracts.
- TrackPayload: useShareImport imports from contracts instead of redefining.
