+++ apps/web/src/stores/types.ts
@@ -1,33 +1,15 @@
 /**
  * Shared domain types for the player stores.
  *
- * The player state was split into `useLibraryStore`, `usePlaybackStore`, and
- * `usePlayerUIStore`, but the `Track`/`RepeatMode` types are referenced by
- * every consumer. Keep them here as the single source of truth so no store
- * owns the canonical definition.
+ * The renderer's `Track` is the canonical `DisplayTrack` from
+ * `@shiranami/contracts` — the display-shaped projection of the DB-mirror
+ * `Track` (nulls collapsed at the `trackMapper` boundary). It's re-exported
+ * here under the `Track` name every store/component already imports, so the
+ * shape lives in one place (contracts) without churning ~118 call sites.
  */
 
-export interface Track {
-  id: string;
-  title: string;
-  artist: string;
-  album: string;
-  duration: number;
-  filePath: string;
-  albumArt?: string;
-  genre?: string | null;
-  year?: number | null;
-  trackNumber?: number | null;
-  discNumber?: number | null;
-  /**
-   * Seed value from the DB. After an in-session toggle, the live value lives
-   * in `useTrackOverlayStore` keyed by track id; read via `useTrack(id)` or
-   * `useMergedLibrary()` rather than off this raw field.
-   */
-  isFavorite?: boolean;
-  playCount?: number;
-  createdAt?: string;
-  updatedAt?: string;
-}
+import type { DisplayTrack } from '@shiranami/contracts';
+
+export type Track = DisplayTrack;
 
 export type RepeatMode = 'off' | 'all' | 'one';
