+++ packages/contracts/src/index.ts
@@ -1,6 +1,7 @@
 // Typed contracts for Shiranami: domain types, IPC channel manifest, and
 // shared API DTOs.
 export * from './domain/track';
+export * from './domain/media';
 export * from './ipc/channels';
 export * from './ipc/debug';
 export * from './ipc/metadata';
